Adds GTFS_DB model in a stand-alone file
